Move parsers to parser package
