+++ src/main/java/seedu/duke/email/command/EmailTagCommand.java
@@ -20,7 +20,7 @@ public class EmailTagCommand extends Command {
      * @param index specific email which tags should be added to
      * @param tags  tags to be added to the email
      */
-    EmailTagCommand(int index, ArrayList<String> tags) {
+    public EmailTagCommand(int index, ArrayList<String> tags) {
         this.index = index;
         this.tags = tags;
     }
