+++ src/main/java/seedu/duke/email/entity/Email.java
@@ -3,9 +3,9 @@ package seedu.duke.email.entity;
 import org.json.JSONArray;
 import org.json.JSONException;
 import org.json.JSONObject;
-import seedu.duke.email.EmailContentParseHelper;
-import seedu.duke.email.EmailFormatParseHelper;
-import seedu.duke.email.EmailStorage;
+import seedu.duke.email.parser.EmailContentParseHelper;
+import seedu.duke.email.parser.EmailFormatParseHelper;
+import seedu.duke.email.storage.EmailStorage;
 import seedu.duke.ui.UI;
 
 import java.io.File;
