+++ src/main/java/seedu/duke/task/command/TaskAddCommand.java
@@ -34,8 +34,8 @@ public class TaskAddCommand extends Command {
      * @param tags     tag associated with the task
      * @param priority priority level of the task
      */
-    TaskAddCommand(Task.TaskType taskType, String name, LocalDateTime time,
-                   String doAfter, ArrayList<String> tags, String priority) {
+    public TaskAddCommand(Task.TaskType taskType, String name, LocalDateTime time,
+                          String doAfter, ArrayList<String> tags, String priority) {
         this.taskType = taskType;
         this.name = name;
         this.time = time;
