+++ src/main/java/seedu/duke/task/command/TaskListCommand.java
@@ -13,7 +13,7 @@ public class TaskListCommand extends Command {
     /**
      * Instantiation of the list command with the target task list.
      */
-    TaskListCommand() {
+    public TaskListCommand() {
 
     }
