+++ src/test/java/seedu/duke/EmailContentParserTest.java
@@ -3,7 +3,7 @@ package seedu.duke;
 import org.junit.jupiter.api.Test;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
-import static seedu.duke.email.EmailContentParseHelper.editDistance;
+import static seedu.duke.email.parser.EmailContentParseHelper.editDistance;
 
 public class EmailContentParserTest {
     @Test
