+++ src/test/java/seedu/duke/EmailTagsTest.java
@@ -5,8 +5,8 @@ import static org.junit.jupiter.api.Assertions.assertFalse;
 import static org.junit.jupiter.api.Assertions.assertTrue;
 
 import seedu.duke.email.EmailList;
-import seedu.duke.email.EmailStorage;
 import seedu.duke.email.EmailTags;
+import seedu.duke.email.storage.EmailStorage;
 
 import java.util.HashMap;
